+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/pyaud/compare/v3.12.1...HEAD)
 ------------------------------------------------------------------------
 ### Added
+- Adds individual file cache
 - Adds `FixAll` plugin
 
 ### Change